requirements.txt
# appnope==0.1.0
# decorator==4.0.9
# docopt==0.6.2
# gnureadline==6.3.3
# ipython==4.1.1
# ipython-genutils==0.1.0
numpy==1.10.4
# path.py==8.1.2
# pexpect==4.0.1
# pickleshare==0.6
# ptyprocess==0.5.1
pyserial==3.0.1
scipy==0.17.0
# simplegeneric==0.8.1
tabulate==0.7.5
# traitlets==4.1.0
urwid==1.3.1
